+++ packages/ui-browse/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @sharelyai/ui-browse
 
+## 0.0.25
+
+### Patch Changes
+
+- update rename chat functions
+- Updated dependencies
+  - @sharelyai/ui-search@0.0.25
+  - @sharelyai/ui-shared@0.0.25
+  - @sharelyai/services@0.0.25
+
+## 0.0.24
+
+### Patch Changes
+
+- Updated dependencies
+  - @sharelyai/services@0.0.24
+  - @sharelyai/ui-search@0.0.24
+  - @sharelyai/ui-shared@0.0.24
+
 ## 0.0.23
 
 ### Patch Changes